+++ changelog.md
@@ -1,19 +1,31 @@
 # betsol-ng-entity-list changelog
 
+## Version 0.0.4
+(25 Oct 2015)
+
+- Improved support for date and datetime field types
+- Added support for links in fields
+- Added support for image field type
+- Added support for sorting parameters
+
+
 ## Version 0.0.3
 (15 Oct 2015)
 
 - Implemented support for deep entity properties
+
 
 ## Version 0.0.2
 (15 Oct 2015)
 
 - Added support for paginator criteria
 
+
 ## Version 0.0.1
 (15 Oct 2015)
 
 - Added support for formatters
+
 
 ## Version 0.0.0
 (17 Sep 2015)